Flytt lagVilkårsvurderingOppfylt til dataGenerator
